+++ src/cpr.cpp
@@ -163,7 +163,7 @@ arma::vec coarsen_theta(unsigned int j, const arma::vec& xi, unsigned int k, con
 }
 
 // [[Rcpp::export]]
-Rcpp::List hat_theta(unsigned int j, const arma::vec& xi, unsigned int k, const arma::vec& theta) {
+Rcpp::List hat_theta(unsigned int j, const arma::vec& xi, unsigned int k, const arma::vec& theta, bool calculate_F, const arma::mat& Sigma) {
   arma::vec xi_sans_j(xi.n_elem - 1);
   for (unsigned int i = 0; i < xi_sans_j.n_elem; ++i) {
     if (i < j) {
@@ -174,12 +174,23 @@ Rcpp::List hat_theta(unsigned int j, const arma::vec& xi, unsigned int k, const 
   }
 
   arma::mat w = W(xi(j), xi_sans_j, k);
+  arma::mat HAT = w * (w.t() * w).i() * w.t();
+  arma::mat IHAT = arma::eye(theta.n_elem, theta.n_elem) - HAT;
+
+  double chisq;
+  if (calculate_F) {
+    arma::mat F = (IHAT * theta).t() * arma::pinv(IHAT * Sigma * IHAT.t()) * (IHAT * theta);
+    chisq = F(0,0);
+  } else {
+    chisq = NA_REAL;
+  }
 
   return (
     Rcpp::List::create(
-        Rcpp::Named("theta") = (w * (w.t() * w).i() * w.t() * theta),
-        Rcpp::Named("d") = theta - (w * (w.t() * w).i() * w.t() * theta),
-        Rcpp::Named("influence") = arma::sum( arma::pow(theta - (w * (w.t() * w).i() * w.t() * theta), 2))
+        Rcpp::Named("theta") = (HAT * theta),
+        Rcpp::Named("d") = IHAT * theta,
+        Rcpp::Named("influence") = arma::norm(IHAT * theta, 2),
+        Rcpp::Named("chisq") = chisq
         )
   );
 }